+++ AlembicImporter/Assets/UTJ/Alembic/Scripts/Exporter/AlembicExporter_impl.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: e5bbd20233d98964fb2c73fa71fe2233
+timeCreated: 1517283688
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
